% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format2}
\alias{format2}
\title{Format numeric values to fixed decimal places}
\usage{
format2(x, digits = 2)
}
\arguments{
\item{x}{Numeric value to format}

\item{digits}{Number of decimal places (default: 2)}
}
\value{
Formatted character string
}
\description{
Format numeric values to fixed decimal places
}
\examples{
format2(3.14159, digits = 2)
# "3.14"
}
