% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{latexify}
\alias{latexify}
\title{Encode text for LaTeX output}
\usage{
latexify(x)
}
\arguments{
\item{x}{Character string to encode}
}
\value{
LaTeX-safe character string
}
\description{
Escapes special LaTeX characters and converts newlines to \\\\
}
\examples{
latexify("100\% sure")
# "100\\\% sure"
}
