% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R, R/mediatr_functions.R
\name{med_diagram_graphviz}
\alias{med_diagram_graphviz}
\title{Create mediation diagram using DiagrammeR/GraphViz}
\usage{
med_diagram_graphviz(
  data,
  height = 0.75,
  width = 2,
  graph_label = "Mediation Diagram",
  node_text_size = 12,
  edge_text_size = 12,
  color = "black",
  ranksep = 0.2,
  minlen = 3
)

med_diagram_graphviz(
  data,
  height = 0.75,
  width = 2,
  graph_label = "Mediation Diagram",
  node_text_size = 12,
  edge_text_size = 12,
  color = "black",
  ranksep = 0.2,
  minlen = 3
)
}
\arguments{
\item{data}{Data frame from med_prepare_diagram_data()}

\item{height}{Node height in inches}

\item{width}{Node width in inches}

\item{graph_label}{Overall diagram label}

\item{node_text_size}{Font size for node labels}

\item{edge_text_size}{Font size for edge labels}

\item{color}{Node and edge color}

\item{ranksep}{Vertical spacing between ranks}

\item{minlen}{Minimum edge length}
}
\value{
GraphViz diagram specification (use with DiagrammeR::grViz())

GraphViz diagram specification
}
\description{
Create mediation diagram using DiagrammeR/GraphViz

Create mediation diagram using DiagrammeR/GraphViz
}
