% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R
\name{med_diagram_plotmat}
\alias{med_diagram_plotmat}
\title{Create mediation diagram using diagram package}
\usage{
med_diagram_plotmat(
  data,
  box_size = 0.15,
  box_prop = 0.5,
  arrow_size = 1.5,
  text_size = 1
)
}
\arguments{
\item{data}{Data frame from med_prepare_diagram_data()}

\item{box_size}{Size of boxes}

\item{box_prop}{Box proportions}

\item{arrow_size}{Size of arrows}

\item{text_size}{Size of text}
}
\value{
Plot object (displays immediately)
}
\description{
Create mediation diagram using diagram package
}
