% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R, R/mediatr_functions.R
\name{med_extract_coefs}
\alias{med_extract_coefs}
\title{Extract formatted coefficients for diagram labels}
\usage{
med_extract_coefs(
  mediation_out,
  model_x_on_m,
  ci = TRUE,
  decimals = 2,
  exp = FALSE
)

med_extract_coefs(
  mediation_out,
  model_x_on_m,
  ci = TRUE,
  decimals = 2,
  exp = FALSE
)
}
\arguments{
\item{mediation_out}{Output from mediation::mediate()}

\item{model_x_on_m}{Linear model of mediator ~ treatment}

\item{ci}{Include confidence intervals in output?}

\item{decimals}{Number of decimal places}

\item{exp}{Exponentiate coefficients?}
}
\value{
Data frame with formatted coefficient strings

Data frame with formatted coefficient strings
}
\description{
Extract formatted coefficients for diagram labels

Extract formatted coefficients for diagram labels
}
