% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{starify}
\alias{starify}
\title{Convert p-values to significance stars}
\usage{
starify(x, cutpoint = 0.05)
}
\arguments{
\item{x}{Numeric p-value}

\item{cutpoint}{Significance threshold (default: 0.05)}
}
\value{
Character string with significance stars
}
\description{
Convert p-values to significance stars
}
