% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrams.R, R/mediatr_functions.R
\name{med_diagram_tikz}
\alias{med_diagram_tikz}
\title{Create mediation diagram in TikZ format for LaTeX}
\usage{
med_diagram_tikz(
  data,
  scale = 0.8,
  box_width = "0.75in",
  box_height = "0.4in",
  caption = "",
  label = "",
  include_figure = TRUE
)

med_diagram_tikz(
  data,
  scale = 0.8,
  box_width = "0.75in",
  box_height = "0.4in",
  caption = "",
  label = "",
  include_figure = TRUE
)
}
\arguments{
\item{data}{Data frame from med_prepare_diagram_data()}

\item{scale}{Overall diagram scale}

\item{box_width}{Width of node boxes}

\item{box_height}{Height of node boxes}

\item{caption}{Figure caption}

\item{label}{LaTeX label for cross-referencing}

\item{include_figure}{Include figure environment?}
}
\value{
TikZ code as string

TikZ code as string
}
\description{
Create mediation diagram in TikZ format for LaTeX

Create mediation diagram in TikZ format for LaTeX
}
